server/requirements.txt
# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0
Flask>=2.0.0
python-dotenv>=0.19.0
streamlink>=6.0.0
eventlet>=0.33.0

# Image and Video Processing
Pillow>=9.0.0
mediapipe>=0.10.0

# Server and API
requests>=2.28.0
websockets>=10.0
python-socketio>=5.7.0
gunicorn>=20.1.0

# Database
SQLAlchemy>=1.4.0
psycopg2-binary>=2.9.0

# Utilities
python-dateutil>=2.8.0